+++ plone/app/widgets/tests/test_dx.py
@@ -786,7 +786,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
     def test_widget_addform_url_on_addform(self):
         from plone.app.widgets.dx import AjaxSelectWidget
         widget = AjaxSelectWidget(self.request)
-        form = Mock()
+        form = Mock(parentForm=None)
         from zope.interface import directlyProvides
         from z3c.form.interfaces import IAddForm
         directlyProvides(form, IAddForm)